# backend/api.yaml
openapi: 3.0.3
info:
  title: Campus Hire
  version: 1.0.0
  description: APIs for Campus Hire.
servers:
  - url: http://localhost:5000
...
paths:
...
  /api/user/registration:
    post:
      tags:
        - Student
      summary: Register Student
      description: Register a new student account.
      requestBody:
        required: true
      responses:
        "201":
          description: Registration successful
        "400":
          description: Invalid request
  /api/user/get-profile:
    get:
      tags:
        - Student
      summary: Get Student Profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Student profile returned
        "404":
          description: Student not found
  /api/user/update-profile:
    put:
      tags:
        - Student
      summary: Update Student Profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
      responses:
        "200":
          description: Profile updated successfully
        "400":
          description: Invalid request
  /api/user/get-active-drives:
    get:
      tags:
        - Student
      summary: Get Active Placement Drives
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of active placement drives
  /api/user/get-drive/{id}:
    get:
      tags:
        - Student
      summary: Get Placement Drive Details
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Placement drive details
        "404":
          description: Drive not found
  /api/user/get-active-companies:
    get:
      tags:
        - Student
      summary: Get Approved Companies
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of approved companies
  /api/user/apply-drive/{drive_id}:
    get:
      tags:
        - Student
      summary: Apply for Placement Drive
      security:
        - bearerAuth: []
      parameters:
        - name: drive_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Application submitted successfully
        "400":
          description: Not eligible or already applied
        "404":
          description: Drive not found
  /api/user/get-history:
    get:
      tags:
        - Student
      summary: Get Application History
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Student application history
  /api/user/export-placement-records:
    get:
      tags:
        - Student
      summary: Export Placement Records
      description: Generates placement report and sends it to the student's email.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Export request submitted
  /api/user/get-company/{id}:
    get:
      tags:
        - Student
      summary: Get Company Details
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Company details with active placement drives
        "404":
          description: Company not found
  /api/user/search:
    get:
      tags:
        - Student
        - Admin
      summary: Global Search
      description: Search students, companies and placement drives depending on the logged-in user's role.
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
          description: Search keyword
      responses:
        "200":
          description: Search results returned successfully
...